+++ server/src/main/java/com/hexacore/tayo/reservation/dto/CreateReservationResponseDto.java
@@ -1,0 +1,11 @@
+package com.hexacore.tayo.reservation.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class CreateReservationResponseDto {
+    private Long reservationId;
+    private Integer fee;
+}
